function/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
sendgrid==6.3.0
azure-servicebus==0.50.2
psycopg2-binary==2.8.6
